Reworked underlining logic to use TextEditorWidget::setExtraSelections() as @Typz suggested.
The new logic seems to work correctly and it looks like it resolved the issue that sometimes the first editor did not underline spelling mistakes.
This update seemed to fix #26 :) I think previously it was updated but the message was remembered, using the new selections all changes with the given Core::Id are replaced by the new selections.