Added logic to specify projects to ignore.
A user can add projects that should not be spell checked. This originated from having to open QtCreator and other large projects many times while working on the plugin and other projects. With this option the plugin will not try to parse those projects, which can take a long time to complete.