Added functionality to sort the Navigation Widget.
Added a menu on the widget where the user can select the sorting order on the navigation widget.
The sorting options are:
- Sort by file name (Default, Asc)
- Sort by total mistakes (Desc)
- Sort by String literal mistakes (Desc)
- Sort by file type (Asc)
The sorting order is fixed and can not be selected by the user. The default was chosen as the one that should make the most sense for such sorting.

The files will always be grouped by internal and external files together. Internal files refer to files part of the active project. In the grouping the files will be sorted according to the user selection.
This closes #33.

Also fixed the display of file names in the case where the file names are too long to fit into the widget.