Fixed issue with trying to parse non-cpp sources.
Parsing non cpp sources (cpp and h files) were not the intended behaviour. A new/different solution should be investigated for parsing these files. There was no guarantee that what the parser did or any actions would have been valid for any other type of file.
This closes #31 but a new issue #36 was created for this.