Added support to parse c and hpp files.
Initial update to add the support for c and hpp files, which were not parsed previously. An investigation is underway to make use of the Qt Creator MIME database to support based on that data and not the file extension.

This update is small enough and does what is needed for now, in the attempt to get binaries out for Qt 3.4.

This should initially close #42, close #39