Avoid 'blocking' the IDE main thread in outputpane
When a file has many errors, the outpane ends up blocking the UI due to
multiple things:
* The non-uniform row height causes very slow layout update when the
selected item changes
* The automatic resizing of the column width

To overcome this, I disabled these two features. The buttons are now
rendered with a simpler, flatter style, which fits in the cell. Position
of each button seems to match approximately with the position of the
underlying text.

The column are now manually resizable, with sensible defaults. The index
column has also been removed (to spare a few pixels), as it does not add
any useful information for the user.

The code was also updated to handle cases with many proposals, to allowg
the layout to overlap the next columns instead of compressing the text
in each button.